requirements.txt
gradio==3.29.0
langchain==0.0.200
neo4j==5.2.0
networkx==3.2
pinecone-client==2.1.0
python-dotenv==1.0.0
langchain_groq
langchain_community